refactor(client): complete with error when manually dismissal
